// src/types/profile.d.ts
export type GithubProfile = {
  login: string;
  avatar_url: string;
  html_url: string;
  repos_url: string;
  company: string;
  blog: string;
  name: string;
  location: any;
  bio: string;
  twitter_username: string;
  public_repos: number;
  followers: number;
  following: number;
  created_at: Date;
};